feat(member): update app scheme and add react-qr-code dependency
