+++ app.config.json
@@ -2,7 +2,7 @@
 	"expo": {
 		"name": "Neuland Next",
 		"slug": "neuland-app-native",
-		"scheme": "neuland",
+		"scheme": "app.neuland",
 		"owner": "neuland-ingolstadt",
 		"version": "0.13.7",
 		"githubUrl": "https://github.com/neuland-ingolstadt/neuland.app-native/",
