+++ src/components/Settings/index.ts
@@ -2,5 +2,6 @@ import Avatar from './Avatar'
 import GradesButton from './GradesButton'
 import InfoBox from './InfoBox'
 import NameBox from './NameBox'
+import NeulandBox from './NeulandBox'
 
-export { Avatar, NameBox, GradesButton, InfoBox }
+export { Avatar, NameBox, GradesButton, InfoBox, NeulandBox }
